Rename gRPC transport settings and move to aux package.
