+++ server/src/main/java/org/opensearch/aux/package-info.java
@@ -1,0 +1,12 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/**
+ * Settings and configuration helpers for optional auxiliary client/server transports provided by installed plugins.
+ */
+package org.opensearch.aux;
