+++ server/src/main/java/org/opensearch/common/settings/ClusterSettings.java
@@ -41,6 +41,7 @@ import org.opensearch.action.search.TransportSearchAction;
 import org.opensearch.action.support.AutoCreateIndex;
 import org.opensearch.action.support.DestructiveOperations;
 import org.opensearch.action.support.replication.TransportReplicationAction;
+import org.opensearch.aux.AuxTransportSettings;
 import org.opensearch.bootstrap.BootstrapSettings;
 import org.opensearch.client.Client;
 import org.opensearch.cluster.ClusterModule;
@@ -111,7 +112,6 @@ import org.opensearch.gateway.remote.RemoteClusterStateCleanupManager;
 import org.opensearch.gateway.remote.RemoteClusterStateService;
 import org.opensearch.gateway.remote.RemoteIndexMetadataManager;
 import org.opensearch.gateway.remote.model.RemoteRoutingTableBlobStore;
-import org.opensearch.grpc.AuxTransportSettings;
 import org.opensearch.http.HttpTransportSettings;
 import org.opensearch.index.IndexModule;
 import org.opensearch.index.IndexSettings;
